requirements.txt
# Aria Personal Agent v7 — Python dependencies
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
requests>=2.31.0
httpx>=0.26.0

# ── Local Inference (replaces Ollama) ──────────────────────────
# Install the correct variant for your hardware:
#
#   CPU only:
#     pip install llama-cpp-python
#
#   NVIDIA GPU (CUDA):
#     CMAKE_ARGS="-DGGML_CUDA=on" pip install llama-cpp-python --upgrade --force-reinstall --no-cache-dir
#
#   Apple Silicon (Metal):
#     CMAKE_ARGS="-DGGML_METAL=on" pip install llama-cpp-python --upgrade --force-reinstall --no-cache-dir
#
#   AMD GPU (ROCm):
#     CMAKE_ARGS="-DGGML_HIPBLAS=on" pip install llama-cpp-python --upgrade --force-reinstall --no-cache-dir
#
# The setup script (setup_v7.py) handles this automatically.
# llama-cpp-python  # installed by setup_v7.py

# ── HuggingFace Hub (model search + download) ──────────────────
huggingface-hub>=0.22.0

# File processing
openpyxl>=3.1.2
python-docx>=1.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.1
pdfminer.six>=20221105

# Multipart / file upload
python-multipart>=0.0.9

# Web search
duckduckgo_search>=5.0.0

# Vector memory (knowledge base)
chromadb>=0.4.22
sentence-transformers>=2.6.0

# Scheduling
apscheduler>=3.10.4

# System monitoring
psutil>=5.9.8

# Timezone
pytz>=2024.1

# Database
# (uses stdlib sqlite3 — no extra install needed)

# Sync (optional — only needed if using Supabase)
supabase>=2.3.0

# Langchain tools (optional, gracefully skipped)
# langchain>=0.1.0
